fetcher/requirements.txt
python-dotenv
aiohttp
pyzmq
pyyaml